LayoutTests/imported/w3c/web-platform-tests/css/css-values/minmax-length-serialize-expected.txt


PASS e.style['letter-spacing'] = "min(1px)" should set the property value 
FAIL e.style['letter-spacing'] = "min(1cm)" should set the property value assert_equals: serialization should be canonical expected "min(1cm)" but got "min(37.79527559055118px)"
FAIL e.style['letter-spacing'] = "min(1mm)" should set the property value assert_equals: serialization should be canonical expected "min(1mm)" but got "min(3.7795275590551185px)"
FAIL e.style['letter-spacing'] = "min(1Q)" should set the property value assert_equals: serialization should be canonical expected "min(1q)" but got "min(0.9448818897637796px)"
FAIL e.style['letter-spacing'] = "min(1in)" should set the property value assert_equals: serialization should be canonical expected "min(1in)" but got "min(96px)"
FAIL e.style['letter-spacing'] = "min(1pc)" should set the property value assert_equals: serialization should be canonical expected "min(1pc)" but got "min(16px)"
FAIL e.style['letter-spacing'] = "min(1pt)" should set the property value assert_equals: serialization should be canonical expected "min(1pt)" but got "min(1.3333333333333333px)"
PASS e.style['letter-spacing'] = "min(1em)" should set the property value 
PASS e.style['letter-spacing'] = "min(1ex)" should set the property value 
PASS e.style['letter-spacing'] = "min(1ch)" should set the property value 
PASS e.style['letter-spacing'] = "min(1rem)" should set the property value 
PASS e.style['letter-spacing'] = "min(1vh)" should set the property value 
PASS e.style['letter-spacing'] = "min(1vw)" should set the property value 
PASS e.style['letter-spacing'] = "min(1vmin)" should set the property value 
PASS e.style['letter-spacing'] = "min(1vmax)" should set the property value 
PASS e.style['letter-spacing'] = "max(1px)" should set the property value 
FAIL e.style['letter-spacing'] = "max(1cm)" should set the property value assert_equals: serialization should be canonical expected "max(1cm)" but got "max(37.79527559055118px)"
FAIL e.style['letter-spacing'] = "max(1mm)" should set the property value assert_equals: serialization should be canonical expected "max(1mm)" but got "max(3.7795275590551185px)"
FAIL e.style['letter-spacing'] = "max(1Q)" should set the property value assert_equals: serialization should be canonical expected "max(1q)" but got "max(0.9448818897637796px)"
FAIL e.style['letter-spacing'] = "max(1in)" should set the property value assert_equals: serialization should be canonical expected "max(1in)" but got "max(96px)"
FAIL e.style['letter-spacing'] = "max(1pc)" should set the property value assert_equals: serialization should be canonical expected "max(1pc)" but got "max(16px)"
FAIL e.style['letter-spacing'] = "max(1pt)" should set the property value assert_equals: serialization should be canonical expected "max(1pt)" but got "max(1.3333333333333333px)"
PASS e.style['letter-spacing'] = "max(1em)" should set the property value 
PASS e.style['letter-spacing'] = "max(1ex)" should set the property value 
PASS e.style['letter-spacing'] = "max(1ch)" should set the property value 
PASS e.style['letter-spacing'] = "max(1rem)" should set the property value 
PASS e.style['letter-spacing'] = "max(1vh)" should set the property value 
PASS e.style['letter-spacing'] = "max(1vw)" should set the property value 
PASS e.style['letter-spacing'] = "max(1vmin)" should set the property value 
PASS e.style['letter-spacing'] = "max(1vmax)" should set the property value 
FAIL e.style['letter-spacing'] = "min(10px, 20px, 30px)" should set the property value assert_equals: serialization should be canonical expected "min(10px, 20px, 30px)" but got "min(10px)"
FAIL e.style['letter-spacing'] = "min(30px, 20px, 10px)" should set the property value assert_equals: serialization should be canonical expected "min(30px, 20px, 10px)" but got "min(10px)"
PASS e.style['letter-spacing'] = "min(20px, 1em, 10vw)" should set the property value 
PASS e.style['letter-spacing'] = "min(10vw, 1em, 20px)" should set the property value 
FAIL e.style['letter-spacing'] = "max(10px, 20px, 30px)" should set the property value assert_equals: serialization should be canonical expected "max(10px, 20px, 30px)" but got "max(30px)"
FAIL e.style['letter-spacing'] = "max(30px, 20px, 10px)" should set the property value assert_equals: serialization should be canonical expected "max(30px, 20px, 10px)" but got "max(30px)"
PASS e.style['letter-spacing'] = "max(20px, 1em, 10vw)" should set the property value 
PASS e.style['letter-spacing'] = "max(10vw, 1em, 20px)" should set the property value 
FAIL e.style['letter-spacing'] = "calc(min(10px) + max(1em) + min(10vw))" should set the property value assert_equals: serialization should be canonical expected "calc(min(10px) + max(1em) + min(10vw))" but got "calc(1em + 10px + 10vw)"
FAIL e.style['letter-spacing'] = "calc(max(1em) + min(10vw) + max(10px))" should set the property value assert_equals: serialization should be canonical expected "calc(max(1em) + min(10vw) + max(10px))" but got "calc(1em + 10px + 10vw)"
FAIL e.style['letter-spacing'] = "calc(10px + min(1em))" should set the property value assert_equals: serialization should be canonical expected "calc(10px + min(1em))" but got "calc(1em + 10px)"
FAIL e.style['letter-spacing'] = "calc(min(1em) + 10px)" should set the property value assert_equals: serialization should be canonical expected "calc(10px + min(1em))" but got "calc(1em + 10px)"
FAIL e.style['letter-spacing'] = "calc(10px + max(1em))" should set the property value assert_equals: serialization should be canonical expected "calc(10px + max(1em))" but got "calc(1em + 10px)"
FAIL e.style['letter-spacing'] = "calc(max(1em) + 10px)" should set the property value assert_equals: serialization should be canonical expected "calc(10px + max(1em))" but got "calc(1em + 10px)"
